MS-375 - Change INT to BIGINT for creating datatables to avoid overflow issues when clients enter phonenumbers

+++ mifosng-provider/src/main/java/org/mifosplatform/infrastructure/dataqueries/service/ReadWriteNonCoreDataServiceImpl.java
@@ -96,7 +96,7 @@ public class ReadWriteNonCoreDataServiceImpl implements ReadWriteNonCoreDataServ
         {
             put("string", "VARCHAR");
             put("checkbox", "VARCHAR");
-            put("number", "INT");
+            put("number", "BIGINT");
             put("boolean", "BIT");
             put("decimal", "DECIMAL");
             put("date", "DATE");
